requirements.txt
google-adk
litellm

# SWE-Bench-Lite
pandas
pyarrow
fastparquet
huggingface_hub